Move LOGISTIC_REGRESSION_ITER to LOGISTIC_REGRESSION
And delete old LOGISTIC_REGRESSION (multiple-local-global version)
