+++ Exareme-Docker/src/mip-algorithms/LOGISTIC_REGRESSION/init/1/global.py
@@ -10,36 +10,24 @@ import numpy as np
 
 sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath(__file__))))) + '/utils/')
 sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath(__file__))))) +
-                '/LOGISTIC_REGRESSION_ITER/')
+                '/LOGISTIC_REGRESSION/')
 
 from algorithm_utils import StateData
-from log_regr_lib import LogRegrIter_Loc2Glob_TD, LogRegrIter_Glob2Loc_TD
+from log_regr_lib import LogRegrInit_Loc2Glob_TD, LogRegrIter_Glob2Loc_TD
 
 
-def logregr_global_iter(global_state, global_in):
-    # Unpack global state
-    n_obs = global_state['n_obs']
-    n_cols = global_state['n_cols']
-    ll_old = global_state['ll']
-    coeff = global_state['coeff']
-    y_val_dict = global_state['y_val_dict']
-    schema_X = global_state['schema_X']
-    schema_Y = global_state['schema_Y']
-    # Unpack global input
-    ll_new, grad, hess = global_in.get_data()
+def logregr_global_init(global_in):
+    n_obs, n_cols, y_val_dict, schema_X, schema_Y = global_in.get_data()
 
-    # Compute new coefficients
-    coeff = np.dot(
-            np.linalg.inv(hess),
-            grad
-    )
-    # Compute delta
-    delta = abs(ll_new - ll_old)
+    # Init vars
+    ll = - 2 * n_obs * np.log(2)
+    coeff = np.zeros(n_cols)
 
     # Pack state and results
-    global_state = StateData(n_obs=n_obs, n_cols=n_cols, ll=ll_new, coeff=coeff, delta=delta,
+    global_state = StateData(n_obs=n_obs, n_cols=n_cols, ll=ll, coeff=coeff,
                              y_val_dict=y_val_dict, schema_X=schema_X, schema_Y=schema_Y)
     global_out = LogRegrIter_Glob2Loc_TD(coeff)
+
     return global_state, global_out
 
 
@@ -48,21 +36,16 @@ def main():
     parser = ArgumentParser()
     parser.add_argument('-cur_state_pkl', required=True,
                         help='Path to the pickle file holding the current state.')
-    parser.add_argument('-prev_state_pkl', required=True,
-                        help='Path to the pickle file holding the previous state.')
     parser.add_argument('-local_step_dbs', required=True,
                         help='Path to db holding local step results.')
     args, unknown = parser.parse_known_args()
     fname_cur_state = path.abspath(args.cur_state_pkl)
-    fname_prev_state = path.abspath(args.prev_state_pkl)
     local_dbs = path.abspath(args.local_step_dbs)
 
-    # Load global state
-    global_state = StateData.load(fname_prev_state).data
     # Load local nodes output
-    local_out = LogRegrIter_Loc2Glob_TD.load(local_dbs)
+    local_out = LogRegrInit_Loc2Glob_TD.load(local_dbs)
     # Run algorithm global step
-    global_state, global_out = logregr_global_iter(global_state=global_state, global_in=local_out)
+    global_state, global_out = logregr_global_init(global_in=local_out)
     # Save global state
     if not os.path.exists(os.path.dirname(fname_cur_state)):
         try:
