+++ Exareme-Docker/src/mip-algorithms/LOGISTIC_REGRESSION/init/1/local.py
@@ -10,8 +10,9 @@ from argparse import ArgumentParser
 
 import numpy as np
 
-sys.path.append(path.dirname(path.dirname(path.dirname(path.abspath(__file__)))) + '/utils/')
-sys.path.append(path.dirname(path.dirname(path.dirname(path.abspath(__file__)))) + '/LOGISTIC_REGRESSION/')
+sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath(__file__))))) + '/utils/')
+sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath(__file__))))) +
+                '/LOGISTIC_REGRESSION/')
 
 from algorithm_utils import StateData
 from log_regr_lib import LogRegrInit_Loc2Glob_TD
