+++ Exareme-Docker/src/mip-algorithms/LOGISTIC_REGRESSION/step/1/global.py
@@ -7,12 +7,12 @@ import os
 import errno
 from argparse import ArgumentParser
 import numpy as np
-import json
 
-sys.path.append(path.dirname(path.dirname(path.dirname(path.abspath(__file__)))) + '/utils/')
-sys.path.append(path.dirname(path.dirname(path.dirname(path.abspath(__file__)))) + '/LOGISTIC_REGRESSION/')
+sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath(__file__))))) + '/utils/')
+sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath(__file__))))) +
+                '/LOGISTIC_REGRESSION/')
 
-from algorithm_utils import StateData, set_algorithms_output_data
+from algorithm_utils import StateData
 from log_regr_lib import LogRegrIter_Loc2Glob_TD, LogRegrIter_Glob2Loc_TD
 
 
@@ -39,10 +39,6 @@ def logregr_global_iter(global_state, global_in):
     # Pack state and results
     global_state = StateData(n_obs=n_obs, n_cols=n_cols, ll=ll_new, coeff=coeff, delta=delta,
                              y_val_dict=y_val_dict, schema_X=schema_X, schema_Y=schema_Y)
-
-    # TODO Dump result to json for testing
-    # global_out = json.dumps({'result': list(coeff)})
-
     global_out = LogRegrIter_Glob2Loc_TD(coeff)
     return global_state, global_out
 
@@ -77,7 +73,6 @@ def main():
     global_state.save(fname=fname_cur_state)
     # Return the algorithm's output
     global_out.transfer()
-    # set_algorithms_output_data(global_out)
 
 
 if __name__ == '__main__':
