--- Exareme-Docker/src/mip-algorithms/LOGISTIC_REGRESSION_ITER/properties.json
@@ -1,50 +1,0 @@
-{
-    "name": "LOGISTIC_REGRESSION_ITER",
-    "desc": "Binary Logistic Regression without regularization using Newton-Raphson method",
-    "type": "python_iterative",
-    "status": "enabled",
-    "parameters": [
-        {
-            "name": "X",
-            "desc": "A list of variables from database. The variable should be Real, Integer. It cannot be empty",
-            "type": "column",
-            "columnValuesSQLType": "real, integer",
-            "columnValuesIsCategorical": "false",
-            "columnValuesNumOfEnumerations": "",
-            "value": "leftaccumbensarea, leftacgganteriorcingulategyrus, leftainsanteriorinsula, rightaccumbensarea, rightacgganteriorcingulategyrus, rightainsanteriorinsula",
-            "valueNotBlank": true,
-            "valueMultiple": true,
-            "valueType": "string"
-        },
-        {
-            "name": "Y",
-            "desc": "A single variable from database. The variable should be Categorical with only two categories (i.e. Boolean). It cannot be empty.",
-            "type": "column",
-            "columnValuesSQLType": "text, integer",
-            "columnValuesIsCategorical": "true",
-            "columnValuesNumOfEnumerations": "2",
-            "value": "gender",
-            "valueNotBlank": true,
-            "valueMultiple": false,
-            "valueType": "string"
-        },
-        {
-            "name": "dataset",
-            "desc": "It contains the names of one or more datasets, in which the algorithm will be executed. It cannot be empty",
-            "type": "dataset",
-            "value": "adni",
-            "valueNotBlank": true,
-            "valueMultiple": true,
-            "valueType": "string"
-        },
-        {
-            "name": "filter",
-            "desc": "",
-            "type": "filter",
-            "value": "",
-            "valueNotBlank": false,
-            "valueMultiple": true,
-            "valueType": "string"
-        }
-    ]
-}
